esc키 기능 추가 (뒤로가기)

+++ android/src/org/synergy/base/Stopwatch.java
@@ -51,7 +51,7 @@ public class Stopwatch {
 	}
 
     private double getClock () {
-        return (double) System.currentTimeMillis () / 1000.0;
+        return System.currentTimeMillis () / 1000.0;
     }
